Back_yrean/requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
python-socketio==5.11.1
python-engineio==4.9.0
boto3==1.34.69
python-dotenv==1.0.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
Flask-CORS==4.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
pytest==8.0.2
pytest-cov==4.1.0 